refactor: Update .gitignore and index.css files
Update .gitignore to ignore the "backend\data\ctg-studies.csv" file.
Update index.css to add styles for the .searchbar-dropdown-clinical class and the option element.

Refactor the server.py file to improve the loading of clinical trials data.
Instead of loading the data every time the function is called, it now checks if the data is already loaded and only loads it if it is not.

Update the ClinicalTrials component to add functionality for downloading search results as a CSV file.
The downloadCSV function generates a timestamp and creates a CSV file with the search results.